使用redis setnx 避免重复消费同一个消息

+++ demo/src/main/java/com/peaknote/demo/service/MessageConsumer.java
@@ -3,12 +3,15 @@ package com.peaknote.demo.service;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.amqp.rabbit.annotation.RabbitListener;
+import org.springframework.data.redis.core.StringRedisTemplate;
 import org.springframework.stereotype.Service;
 
 import com.peaknote.demo.model.TranscriptInfo;
 
 import static com.peaknote.demo.config.RabbitMQConfig.EVENT_QUEUE;
 import static com.peaknote.demo.config.RabbitMQConfig.TRANSCRIPT_QUEUE;
+
+import java.util.concurrent.TimeUnit;
 
 @Service
 @RequiredArgsConstructor
@@ -20,6 +23,7 @@ public class MessageConsumer {
     private final GraphEventService graphEventService;
     private final TranscriptService transcriptService;
     private final String accessToken;
+    private final StringRedisTemplate stringRedisTemplate;
 
     @RabbitListener(queues = EVENT_QUEUE, containerFactory = "rabbitListenerContainerFactory")
     public void handleEventMessage(String payload) {
@@ -27,6 +31,14 @@ public class MessageConsumer {
             log.info("✅ 消费 Event 消息");
             String userId = payloadParserService.extractUserIdFromEventPayload(payload);
             String eventId = payloadParserService.extractEventIdFromEventPayload(payload);
+            String type = payloadParserService.extractChangeType(payload);
+            String key = eventId + ":" + type;
+            // Redis 去重
+            Boolean success = stringRedisTemplate.opsForValue().setIfAbsent(key, "1", 5, TimeUnit.MINUTES);
+            if (Boolean.FALSE.equals(success)) {
+                log.warn("⚠️ 重复消息，已忽略: {}", key);
+                return;
+            }
             boolean isRecurring = (graphService.getUserEvent(userId, eventId).recurrence != null);
             graphEventService.processEvent(userId, eventId, isRecurring);
         } catch (Exception e) {
@@ -40,6 +52,16 @@ public class MessageConsumer {
             log.info("✅ 消费 Transcript 消息");
             TranscriptInfo transcriptInfo = payloadParserService.parseTranscriptInfo(payload);
             String subscriptionId = payloadParserService.extractSubscriptionId(payload);
+            String type = payloadParserService.extractChangeType(payload);
+
+            String key = subscriptionId + ":" + type;
+
+            // Redis 去重
+            Boolean success = stringRedisTemplate.opsForValue().setIfAbsent(key, "1", 5, TimeUnit.MINUTES);
+            if (Boolean.FALSE.equals(success)) {
+                log.warn("⚠️ 重复消息，已忽略: {}", key);
+                return;
+            }
 
             if (subscriptionId != null) {
                 graphService.deleteSubscription(subscriptionId);
